requirements.txt

requests == 2.20.0
google-cloud-secret-manager==2.3.0
google-cloud-storage==1.36.1
PyYAML==5.4.1
google-cloud-bigquery==2.11.0